[Update]敵の処理の構造を大幅に変更、既に敵が配置されていたステージの改修

+++ AMAZURU/Assets/Hara/Scripts/Enemy/EnemyData.cs
@@ -44,5 +44,5 @@ public class EnemyData : ScriptableObject
     public bool UseDefaultSetting = true;
 
     // 乾燥ブロックの設置高さ
-    public float BlockSetPosY = 1.0f;
+    public int BlockSetPosY = 1;
 }
